test: Add unit tests for CommonModels class
- Created `tests/commonmodels_test.cpp` to verify `CommonModels::update` logic.
- Isolated stubs for `WinOcr` and `Translator` static methods into `tests/mock_commonmodels_deps.cpp` to prevent ODR violations and global namespace pollution across the test executable.
- Added `tests/mock_commonmodels_deps.cpp` to `tests.pro` `SOURCES`.
- Appended `src/ocr` and `src/translate` to `tests.pro` `INCLUDEPATH` to allow `commonmodels.cpp` to resolve dependencies without needing to alter production source `#include` statements.
- Restored `src/commonmodels.cpp` to its original state.
- Set `_SILENCE_EXPERIMENTAL_COROUTINE_DEPRECATION_WARNINGS` in `tests.pro` for Windows/MSVC compilation to mirror the main `screen-translator.pro` behavior and fix Windows CI breakage related to `<experimental/coroutine>`.
